- changed password hashing function - removed ExceptionCollector - a few renames of e_logger to e_profiler
