+++ flask_monitoringdashboard/core/exceptions/scoped_exception_collector.py
@@ -1,6 +1,8 @@
 from typing import Union
 import copy
 
+from sqlalchemy.orm import Session
+
 
 class ScopedExceptionCollector:
     """
@@ -19,6 +21,31 @@ class ScopedExceptionCollector:
     def set_uncaught_exc(self, e: BaseException):
         e_copy = _get_copy_of_exception(e)
         self.uncaught_exception = e_copy
+
+    def save_to_db(self, request_id: int, session: Session):
+
+        from flask_monitoringdashboard.database.exception_occurrence import (
+            save_exception_occurence_to_db,
+        )
+
+        """
+        Iterates over all the user captured exceptions and also a possible uncaught one, and saves each exception to the DB
+        """
+        for e in self.user_captured_exceptions:
+            save_exception_occurence_to_db(
+                request_id, session, e, type(e), e.__traceback__, True
+            )
+
+        e = self.uncaught_exception
+        if e is not None:
+            if e.__traceback__ is not None:
+                # We have to choose the next frame as else it will include the evaluate function from measurement.py in the traceback
+                # where it was temporaritly captured for logging by the ScopedExceptionCollector, before getting reraised later
+                e = e.with_traceback(e.__traceback__.tb_next)
+
+            save_exception_occurence_to_db(
+                request_id, session, e, type(e), e.__traceback__, False
+            )
 
 
 def _get_copy_of_exception(e: BaseException):
