+++ flask_monitoringdashboard/core/profiler/outlier_profiler.py
@@ -5,11 +5,9 @@ import traceback
 import psutil
 from flask import request
 
-from flask_monitoringdashboard import config
+from flask_monitoringdashboard import config, ScopedExceptionCollector
 from flask_monitoringdashboard.core.cache import update_duration_cache, get_avg_endpoint
-from flask_monitoringdashboard.core.exceptions.exception_collector import (
-    ExceptionCollector,
-)
+
 from flask_monitoringdashboard.core.logger import log
 from flask_monitoringdashboard.database import session_scope
 from flask_monitoringdashboard.database.outlier import add_outlier
@@ -66,7 +64,7 @@ class OutlierProfiler(threading.Thread):
             self._cpu_percent = str(psutil.cpu_percent(interval=None, percpu=True))
             self._memory = str(psutil.virtual_memory())
 
-    def stop(self, duration, status_code, e_logger: ExceptionCollector):
+    def stop(self, duration, status_code, e_collector: ScopedExceptionCollector):
         self._exit.set()
         update_duration_cache(
             endpoint_name=self._endpoint.name, duration=duration * 1000
@@ -80,7 +78,7 @@ class OutlierProfiler(threading.Thread):
                 group_by=self._group_by,
                 status_code=status_code,
             )
-            e_logger.save_to_db(request_id, session)
+            e_collector.save_to_db(request_id, session)
             if self._memory:
                 add_outlier(
                     session,
