+++ flask_monitoringdashboard/core/profiler/performance_profiler.py
@@ -1,11 +1,10 @@
 from typing import Union
+
+from flask_monitoringdashboard import ScopedExceptionCollector
 from flask_monitoringdashboard.core.cache import update_duration_cache
 from flask_monitoringdashboard.core.profiler.base_profiler import BaseProfiler
 from flask_monitoringdashboard.database import session_scope
 from flask_monitoringdashboard.database.request import add_request
-from flask_monitoringdashboard.core.exceptions.exception_collector import (
-    ExceptionCollector,
-)
 
 
 class PerformanceProfiler(BaseProfiler):
@@ -20,7 +19,7 @@ class PerformanceProfiler(BaseProfiler):
         ip,
         duration,
         group_by,
-        e_logger: ExceptionCollector,
+        e_collector: ScopedExceptionCollector,
         status_code=200,
     ):
         super(PerformanceProfiler, self).__init__(endpoint)
@@ -29,7 +28,7 @@ class PerformanceProfiler(BaseProfiler):
         self._endpoint = endpoint
         self._group_by = group_by
         self._status_code = status_code
-        self.e_logger: ExceptionCollector = e_logger
+        self.e_logger: ScopedExceptionCollector = e_collector
 
     def run(self):
         update_duration_cache(
