+++ flask_monitoringdashboard/database/__init__.py
@@ -52,7 +52,8 @@ class User(Base):
     """False for guest permissions (only view access). True for admin permissions."""
 
     def set_password(self, password):
-        self.password_hash = generate_password_hash(password)
+        # Using the pbkdf2 because it scrypt won't work on a Mac M1 with Sequoia
+        self.password_hash = generate_password_hash(password, "pbkdf2")
 
     def check_password(self, password):
         return check_password_hash(self.password_hash, password)
@@ -249,7 +250,9 @@ class StackTraceSnapshot(Base):
     id = Column(Integer, primary_key=True)
     hash = Column(String(64), nullable=False, unique=True)
 
-    exception_stack_lines = relationship("ExceptionStackLine", back_populates="stack_trace_snapshot")
+    exception_stack_lines = relationship(
+        "ExceptionStackLine", back_populates="stack_trace_snapshot"
+    )
 
 
 class ExceptionType(Base):
@@ -362,7 +365,9 @@ class ExceptionStackLine(Base):
     stack_trace_snapshot_id = Column(
         Integer, ForeignKey(StackTraceSnapshot.id), primary_key=True
     )
-    stack_trace_snapshot = relationship(StackTraceSnapshot, back_populates="exception_stack_lines")
+    stack_trace_snapshot = relationship(
+        StackTraceSnapshot, back_populates="exception_stack_lines"
+    )
     """Stack trace that belongs to this exc_stack_line."""
 
     exception_frame_id = Column(
