+++ flask_monitoringdashboard/database/exception_occurrence.py
@@ -2,9 +2,18 @@
 Contains all functions that access an ExceptionOccurrence object.
 """
 
+from types import TracebackType
 from typing import Union
 from sqlalchemy import func, desc
 from sqlalchemy.orm import Session
+
+from flask_monitoringdashboard.core.exceptions.stack_frame_parsing import (
+    get_function_definition_from_frame,
+)
+from flask_monitoringdashboard.core.exceptions.stack_trace_hashing import (
+    hash_stack_trace,
+)
+
 from flask_monitoringdashboard.database import (
     ExceptionOccurrence,
     Request,
@@ -15,6 +24,22 @@ from flask_monitoringdashboard.database import (
 from flask_monitoringdashboard.core.database_pruning import (
     delete_entries_unreferenced_by_exception_occurrence,
 )
+from flask_monitoringdashboard.database.exception_frame import add_exception_frame
+from flask_monitoringdashboard.database.exception_message import add_exception_message
+from flask_monitoringdashboard.database.exception_stack_line import (
+    add_exception_stack_line,
+)
+from flask_monitoringdashboard.database.exception_type import add_exception_type
+from flask_monitoringdashboard.database.file_path import add_file_path
+from flask_monitoringdashboard.database.function_definition import (
+    add_function_definition,
+)
+from flask_monitoringdashboard.database.function_location import add_function_location
+from flask_monitoringdashboard.database.stack_trace_snapshot import (
+    get_stack_trace_by_hash,
+    add_stack_trace_snapshot,
+)
+
 
 def add_exception_occurrence(
     session: Session,
@@ -162,3 +187,59 @@ def get_exceptions_with_timestamps_and_stack_trace_id(
         .all()
     )
     return result
+
+
+def save_exception_occurence_to_db(
+    request_id: int,
+    session: Session,
+    exc: BaseException,
+    typ: type[BaseException],
+    tb: Union[TracebackType, None],
+    is_user_captured: bool,
+):
+    """
+    Save exception occurence to DB
+    """
+    hashed_trace = hash_stack_trace(exc, tb)
+    existing_trace = get_stack_trace_by_hash(session, hashed_trace)
+
+    if existing_trace:
+        trace_id = int(existing_trace.id)
+    else:
+        trace_id = add_stack_trace_snapshot(session, hashed_trace)
+        idx = 0
+        while tb:
+            # iterate over traceback-type objects
+            # i.e. the object representation of the following traceback
+            # Traceback (most recent call last):
+            #   File "example.py", line 9, in <module>
+            #     calculate()
+            #   File "example.py", line 6, in calculate
+            #     return divide(10, 0)
+            #   File "example.py", line 2, in divide
+            #     return a / b
+            # ZeroDivisionError: division by zero
+            f_def = get_function_definition_from_frame(tb.tb_frame)
+            function_id = add_function_definition(session, f_def)
+            file_path = add_file_path(session, tb.tb_frame.f_code.co_filename)
+            f_location_id = add_function_location(
+                session,
+                file_path,
+                function_id,
+                tb.tb_frame.f_code.co_firstlineno,
+            )
+            frame_id = add_exception_frame(session, f_location_id, tb.tb_lineno)
+            add_exception_stack_line(
+                session,
+                trace_id,
+                frame_id,
+                idx,
+            )
+            tb = tb.tb_next
+            idx += 1
+
+    exc_msg_id = add_exception_message(session, str(exc))
+    exc_type_id = add_exception_type(session, typ.__name__)
+    add_exception_occurrence(
+        session, request_id, trace_id, exc_type_id, exc_msg_id, is_user_captured
+    )
